Add initial support for Query Objects
A new main component of Rectify is the Query Object. It allows a
database query to be encapsulated into a class. This allows logic that
is specific to the query as well as the query itself to be removed from
the ActiveRecord model as another strategy to reducing their size.

Query Objects need to be derived from `Rectify::Query` and must
implement the `query` method that returns either an
`ActiveRecord::Relation` or an array (via
`ActiveRecord::Quering#find_by_sql`)

This commit also adds a simple way to stub Query Objects with the
`stub_query` RSpec helper.

See the readme for full details of how to use Query Objects in a Rails
app.