Add initial experimental database stats for RSpec
This is an experimental feature that will allow developers to get stats
about the number and time taken but database queries in their specs.

To enable, the developer will need to `require "rectify/rspec"` in
their spec/rails_helper.rb file and then add the following:

```
Rectify::RSpec::DatabaseReporter.enable
```

That will register the Rectify database reporter with RSpec and will
start counting queries. Once the suite is complete, a summary is
displayed.