Add support for `Query#merge` and `Query.merge`
After some discussion we decided that some users would prefer a
slightly different way to compose query objects rather than using the
pipe operator.

The first option is to use `Query#merge` which is just an alias to `#|`
and you use it like this:

```
FirstQuery.new(args).merge(SecondQuery.new)
```

We also considered the case when users want to composed a longer list
of queries and would like to format their code over multiple lines. For
that use case we now provide the class method `Query.merge` that takes a
list of queries and merges them all together:

```
combined = Rectify::Query.merge(
  FirstQuery.new(args),
  SecondQuery.new,
  AnotherQuery.new
)
```