Populate Form Objects with associated models
This improves the `Rectify::Form.from_model` method. Before this commit, the form object was only populated by whatever the model returned from it's `#attributes` method. This problem with this is that it only returns attributes and not associations like `has_many` or `belongs_to`. In these cases, even if the form had matching attributes for these, they wouldn't be populated from the model when `.from_model` was called.

This commit changes how `.from_model` works. Now what happens is that the attributes of the form object are iterated over and if the model responds to a message of the same name then the form attribute will be populated. This means that not only will normal attributes, `belongs_to` and `has_many` associations populate the form object but also if the model has a method with a matching form attribute name, it will also be populated.