Add support for indexed arrays in form objects
For some Rails forms built with a form builder, the params hash
contains an indexed hash rather than an array. For example, Rectify
already supported the following params hash when using the
`.from_params` method:

```
{
  "user" => {
    "contacts" => [
      { "name" => "Amber",   "number" => "123" },
      { "name" => "Megan",   "number" => "456" },
      { "name" => "Charlie", "number" => "789" }
    ]
  }
}
```

However, when posting from a Rails form, these array elements can come
through as an indexed hash, for example:

```
{
  "user" => {
    "contacts" => {
      "0" => { "name" => "Amber",   "number" => "123" },
      "1" => { "name" => "Megan",   "number" => "456" },
      "2" => { "name" => "Charlie", "number" => "789" }
    }
  }
}
```

Using Rectify form objects with this structure and trying to populate
an `Array` of nested form objects would result in the error:

```
Expected ["0", {"name"=>"Amber", "number"=>"123"}] to respond to #to_hash
```

To fix this, we now convert a hash in this form to the one above before
passing it into the form object constructor and allowing Virtus to
populate the fields. This fix isn't a deep fix at present and so this
may need to be looked at when/if the situation arises.