Command handler can use caller private methods
There are a number of private methods that it would be useful to access
from within a command handler block when used from inside a Rails
controller.

One reported case is #20 where the reporter would like to access the
cookies collection. This is a private Rails controller method that is
normally accessible within an action.

The handler blocks delegate calls to the caller using method missing.
In that method with were checking first to confirm if the caller could
respond to the message but only included public methods. We now also
check private methods and if supported by the caller we use `#send` to
forward the message onto the caller (public or private methods).

This should fix #20 and also fix #15.