Bump gems and fix issues with Rails 5.2
Rectify works with Rails 4.1 and above but several gems in
Gemfile.lock have vulnerabilities and so this commit bumps them all to
their latest versions.

With that comes the bump to Rails 5.2 and although Rectify continues to
work, we needed to make a few changes to allow development and CI builds
to pass.